Added nGunners field to SmallCraft.
The rules aren't really clear about the effects of adding more crew than
required, but the canon designs typically do so and while most of them
use the minimum number of gunners, there are some with extra. So to
reflect the canon data we need to be able to set the actual number of
gunners.