Munition choice panel now remembers ammo selection
With the advent of caseless ammo, we have munitions which can have more than
the original number of shots.  This exposes a problem with the munition choice
panel.  Namely, it doesn't actually change the number of shots selected when
a munition is changed, instead it will just select the max # of shots if  the
current seleciton is higher than what's available.

I made it so now the panel tracks if the # of shots has been changed, and if
it hasn't, the max is always selected.  This why, switching beetween caseless
or precision will work as intended.