Partial fix for issue #530.
This partially fixes the issue for #530. The system will try to find the help file in where MM expects it, if it can't find it, it assumes we're running MM in MHQ, so it looks for the help file where MHQ expects it.

This is kind of a kludge, and there are more than likely better ways to fix this, but this should work for now.