Added SharedConfiguration class to handle configuration values shared across applications.
This commit takes a radical different direction that the last one. Instead of hard coding a couple of different locations to check for the readme.txt file, I've added a singleton class that can be used to get configuration values from a shared properties file in mmconf. This file is appropriately called, shared.properties.