Allow units to pass under bridges.
Moves under bridges were being marked illegal if climb mode is on and
they are not entering from a road hex, but those with an elevation low
enough to get under the bridge should just ignore it.

Fixes #521: Submarine cannot pass under bridges in a turn when their
elevation changes.