Fix client invalid attack packets during sim play
During a simultaneous phase, it can be possible for the client to select
a unit that is already done.  This happens if the player hits done on a unit
and sends it to the server, but the server is already processing a the move
for another players unit.  The server will issue a turn change when the prev
unit is done, which will cause the first client to fire a turn change event
but because the first client hasn't received a response from the server about
the turn it sent, it can end up reselecting the entity it just moved.

What this change does it make it so during simultaneous phases, the client will
only respond to turn changes that were initiated by the local player.  This
means that each time the client "dones" it will get prompted for a new turn
once the server has processed that turn, but will ignore turn changes from
any other player.