Ensure Bot calculates correct # of turn, sim phase
In simultaneous phases, everytime a player takes a turn, it signals the bot
client to spawn a new CalculateBotTurn thread.  This ultimately means that
the bot will calculate a number of turns equals to the number of units, not
the number of units the bot controls.  This can ultimately lead to odd
situations.  A previous fix for this was no effective.

This fix makes it so that the bot will only respond to turn changes that
resulted from the bot making a turn.