Adjust source of event
I found that ToolTipManager$AccessibilityKeyListener assumes that the event
source is a JComponent.  With the way we were passing forward key events,
this could cause a ClassCastException.  I don't see any reason why we can't
fake the source of the event.  This fixes a very rare exception without
causing further problems (if it does cause further problems we can reverse it).