Board Editor: add size restrictions to hex canvas
This is important for large texture tilesets.  As it is, the image will
be draw as large as it can.  These changes at least limit it to a hex size
although it's a square and not hex shape.  I fiddled around with it, but for
some inexplicable reason the code in BoardView1 that we use wasn't working
here.