Removed getNextEntityNum(int) from Game and IGame.
This was only a convenience method used by bots that called
getNextEntityNum(GameTurn,int) with the Game object's turn but this is
incorrect when using simultaneous phases and the bot's turn differed
from the Game's.

This was previously fixed for Princess, but this also fixes it for
TestBot as well as removing the method that should not be used.