Reworked MechView to be able to output html or plain text.
The MechTextView class (used by MML to export as text) appears to have
started as a copy of MechView but has not been kept updated and updating
two classes with every change in the preview format is more work than
necessary. This combines them into a single class which also
standardizes the presentation with a set of utility classes that should
make future updates easier.