Take intro date into account when calculating era availability.
Previously calculated era availability codes for units using only
components. In the case of a unit that uses, for instance, only
components that are available in the Star League era but has an intro
date in the Clan Invasion era, the availability should be X for SL and
SW eras rather than the availability of the rarest component.