Data: Warship Data Files
This is the first part of a data overhaul for the Large Space Craft.
This removes all the old WS files and structure and replaces them with
an updated structure.

These files have been built in MML with rules from SO Second Printing,
and additional information received from official sources.