Data - Final Changes
This should be the final set of data changes for Large Craft.  This
includes the updated name changes.  For those following the data
releases some of the name changes are inline with the MUL, other name
changes should be considered accurate even if it conflicts with the MUL.

The folder structure has also been re-ordered and will still in this new
format.