Introducing BuildingSection
Note this is an interim commit where the hex-specific building information
is duplicated in both the "old" maps and the new BuildingSection.
This is intended for verification only and the "old" stuff will be removed
in a subsequent commit.

Also, note that build.xml has been updated: when compiling tests, all
resources are now copied to classes/ (instead of  only .xml files).
This is needed in order to run the new tests included in this commit.