Apply candidate limits correctly in noSoLazyPathFinder.
Previously it was adding a MovePath to the candidates queue then
checking against a max size and if over 100 removing the index of the
last item. PriorityQueue does not have a method to remove by index, only
by object reference, so it was doing nothing and it's more efficient to
check before adding it in the first place.