Restrict one-shot weapons to their own bin when configuring ammo.
The previous behavior is to get a list of all compatible ammo bins for
the weapon and link to the first in the list. This resulted in multiple
types of one-shot launchers all sharing the same shot of ammo and could
also allow non one-shot launchers to be linked to one shot ammo if both
varieties of launcher were on the same unit.