Fix Half Light Machine Gun Ammo(IS)
https://github.com/MegaMek/megameklab/issues/314#issue-456358053

Fix for issue #314 of MegaMekLab: IS version of Light Machine Gun Ammo[Half] is not available for Inner Sphere Tech Base

Change setClanAdvancement to setISAdvancement and make it available for IS, not Clans, and also change the year. For now the year is same as Clans half Light Machine Gun Ammo, not IS Full Light Machine Gun Ammo.