Remove tech base check from protmech armor enum lookup.
All protomech armor has a Clan tech base. Since standard uses the MiscType
corresponding to T_STANDARD, and it has an tech level of Intro, it can't make
the Clan/IS distinction and when the armor is set to standard in MML it is
treated as IS and fails to match. The tech base match was kept when the code
was copied from other unit types and modified, but is not necessary for
protomechs.