Move all round report formatting to the Report class
Rename Report.getText to getHtml, since that is what it produces. Move
HTML tag stripping code from AccessibilityWindow to new getText method
in that class. Move the reformatting that happens in
Client.receiveReport to the getHtml method, so enhanced round report
markup can be inserted when the HTML is first generated.