Fixes #3699 Deadfire LRMs and SRMs have incorrect range brackets
With lots of help from Juliez (who has also reviewed and tested) I've
got Dead Fire missiles implemented with the correct ranges and updates
to display. One area I noticed we are missing is correct BV for them.
But it is one of the weapons where every Ammo size has a different BV.
I've added a TODO for this.