Data: New Sprites and start of BA project
The new sprites is pretty straight forward. The BA project will be a
longish project to add the new 4,5,6 unit squads that the MUL has added.
As part of this they will be getting updated sprites and fluff text
where applicable. I'm only going to have one history entry as this will
be complete before next dev release.