Add Logic To Filter On Weapon Class
Adds a new FT called WeaponClassFT that is adjacent to EquipmentFT.
Additionally adds an Integer member to ExpNode. This is set to null when ExpNode is constructed with a String (only constructor before).
But the new constructor (which takes an Integer instead of a String), sets the weaponClass.
If the ExpNode has a non-null weaponClass, we know it's a request for a filter on weaponClass instead of general equipment.
Logic is then added to filter units based on weapon class (which was added to the MechSummary in a previous commit).