Fix for #1227: ArrayIndexOutOfBoundsException when loading Superheavy unit that contains variable-size components.
There is already code to divide the required criticals in this situation
1) in the Unallocated Equipment pane,
2) within Mech.java

But when loading a Variable Size component from an MTF file the code
simply uses the SIZE value as the # of required crits.  This is
accounted for properly in other areas of the code but not when loading
crits from a file.