Fix issue caused by illegal, empty <transporters> blocks.
It appears that we have been writing empty <transporters> blocks
into every vehicle for the past 6 or so years.  These blocks should
not exist in, e.g., combat vehicles, but apparently went unnoticed
due to loose handling of the relevant block when empty.

This change adds a new test prior to processing these empty blocks
on file load, a function to support this test, and a conditional
around writing the <transporters> block to prevent writing empty
blocks in every .BLK file.