Actual fix for #4359
Add the (-2) Flak modifier to direct Flak artillery / artillery cannon
attacks.

This appears to make close misses with Flak artillery more effective, as
drift is lower when the target number is lower but the roll misses.

Tested using Sniper and LT with standard and Cluster rounds vs both
VTOL and ASF targets.

close #4359