Fix parser bug when loading tripod AES
The file parser for loading 'Mech designs previously failed to load
tripods with AES, as it didn't expect that to be located in the center
leg. This has been fixed by adding that location to the list of checks.

I initially assumed that this would impact quadrupeds as well. Since
they don't seem to have their own leg locations, I'm guessing they reuse
the arm locations.