A simple test that checks canonicity is working as intended
Rewrote as requested, also changed the other tests in EntityTest to use files in test resources.

Checks if a known canon mech is understood by MM to be canon. Fails if it isn't. Note that the test should actually be failing at the moment, see https://github.com/MegaMek/megamek/issues/5913 .