Added `TotalGenericBattleValue` Tag to MUL Export
Refactored the `saveTo` method in `EntityListFile` to include an optional generic battle value parameter. Added calculation and integration of generic battle value in `BriefingTab` when saving entity lists to files.

This is hidden in MUL export as GBV is not meant to be user-facing. We include it here as a debugging tool.